requirements.txt
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.20,<0.1.0
langchain-google-genai>=0.0.9,<0.1.0
google-generativeai>=0.4.0,<0.5.0
chromadb>=0.4.22,<0.5.0
pypdf>=3.0.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
streamlit>=1.30.0,<2.0.0